+++ inventory/src/main/java/com/lab/inventory/data/entity/TransactionDetailId.java
@@ -13,7 +13,7 @@ import java.io.Serializable;
 @NoArgsConstructor
 @EqualsAndHashCode
 public class TransactionDetailId implements Serializable {
-    private Long transaction_id;
+    private Integer transaction_id;
     private Integer seq;
 
 }
